[BugFix] Fix wrong validation of 'exec_mem_limit' (#13796)
(cherry picked from commit b4989c47d3858074e98bf46d6964d50401223807)

+++ fe/fe-core/src/main/java/com/starrocks/qe/SessionVariable.java
@@ -323,7 +323,7 @@ public class SessionVariable implements Serializable, Writable, Cloneable {
 
     // Limitations
     // mem limit can't smaller than bufferpool's default page size
-    public static final int MIN_EXEC_MEM_LIMIT = 2097152;
+    public static final long MIN_EXEC_MEM_LIMIT = 2097152;
     // query timeout cannot greater than one month
     public static final int MAX_QUERY_TIMEOUT = 259200;
